+++ frontend/src/metabase/querying/filters/hooks/use-time-filter/utils.ts
@@ -1,8 +1,10 @@
 import dayjs from "dayjs";
 
 import { isNotNull } from "metabase/lib/types";
-import { getDefaultAvailableOperator } from "metabase/querying/filters/utils";
-import { getAvailableOperatorOptions } from "metabase/querying/filters/utils/operators";
+import {
+  getAvailableOperatorOptions,
+  getDefaultAvailableOperator,
+} from "metabase/querying/filters/utils/operators";
 import * as Lib from "metabase-lib";
 
 import { OPERATOR_OPTIONS } from "./constants";
