+++ frontend/src/metabase/querying/filters/utils.ts
@@ -1,7 +1,5 @@
 import type * as Lib from "metabase-lib";
 
-import type { FilterOperatorOption } from "./types";
-
 export function getGroupName(
   groupInfo: Lib.ColumnGroupDisplayInfo,
   stageIndex: number,
@@ -9,14 +7,4 @@ export function getGroupName(
   return groupInfo.isMainGroup && stageIndex > 1
     ? `${groupInfo.displayName} (${stageIndex})`
     : groupInfo.displayName;
-}
-
-export function getDefaultAvailableOperator<T extends Lib.FilterOperatorName>(
-  options: FilterOperatorOption<T>[],
-  desiredOperator?: T,
-): T {
-  return (
-    options.find(option => option.operator === desiredOperator)?.operator ??
-    options[0].operator
-  );
 }
